users/views.py
# D:\UVBFL3\ProjetTutore\users\views.py
from rest_framework.views import APIView
from rest_framework.permissions import IsAuthenticated

from rest_framework import generics, status
from rest_framework.response import Response
from rest_framework.permissions import AllowAny
from django.contrib.auth import authenticate
from rest_framework_simplejwt.tokens import RefreshToken

from .models import User

# Import des serializers corrects
from .serializers import (
    UserSerializer,
    RegisterSerializer,
    LoginSerializer
)


# -------------------------------------------------------
#  REGISTER (INSCRIPTION)
# -------------------------------------------------------
class RegisterView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = RegisterSerializer
    permission_classes = [AllowAny]

    def create(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        user = serializer.save()

        return Response({
            "message": "Utilisateur créé avec succès",
            "user": UserSerializer(user).data
        }, status=status.HTTP_201_CREATED)



# -------------------------------------------------------
#  LOGIN (CONNEXION)
# -------------------------------------------------------
class LoginView(generics.GenericAPIView):
    serializer_class = LoginSerializer
    permission_classes = [AllowAny]

    def post(self, request):
        email = request.data.get("email")
        password = request.data.get("password")

        user = authenticate(email=email, password=password)

        if not user:
            return Response({"error": "Identifiants invalides"}, status=400)

        refresh = RefreshToken.for_user(user)

        return Response({
            "access": str(refresh.access_token),
            "refresh": str(refresh),
            "user": UserSerializer(user).data
        })

class ProfileView(APIView):
    permission_classes = [IsAuthenticated]

    def get(self, request):
        user = request.user
        return Response({
            "id": user.id,
            "username": user.username,
            "email": user.email,
        })