requirements.txt
asgiref==3.10.0
attrs==25.4.0
brotli==1.2.0
cffi==2.0.0
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
fonttools==4.60.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.2.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
